# src/go/web/public/docs/openapi.yml
openapi: 3.0.0
info:
  description: phenix API
  version: 0.0.0
  title: phenix
  termsOfService: http://swagger.io/terms/
  contact:
    url: https://github.com/sandialabs/sceptre-phenix
  license:
    name: GNU General Public License v3.0
...
paths:
  "/configs":
    get:
      tags:
        - Configs
      summary: Get all phenix configs
      description: ""
      operationId: getConfigs
      parameters:
        - name: kind
          in: query
          description: limit configs to specified kind (defaults to all)
          required: false
          schema:
            type: string
            enum:
              - all
              - topology
              - scenario
              - experiment
              - image
              - user
              - role
            default: all
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Configs"
    post:
      tags:
        - Configs
      summary: Create new phenix config
      description: ""
      operationId: postConfigs
      requestBody:
        description: phenix config creation parameters
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Config"
          application/x-yaml:
            schema:
              $ref: "#/components/schemas/Config"
          multipart/form-data:
            schema:
              type: object
              properties:
                fileupload:
                  type: string
                  format: binary
      responses:
        "201":
          description: successful config creation
          headers:
            Location:
              schema:
                description: location of newly created config
                type: string
                format: uri
  "/configs/{kind}/{name}":
    get:
      tags:
        - Configs
      summary: Get existing phenix config
      description: ""
      operationId: getConfigsKindName
      parameters:
        - name: kind
          in: path
          description: kind of phenix config to get
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: name of phenix config to get
          required: true
          schema:
            type: string
        - name: Accept
          in: header
          description: content format for response
          required: false
          schema:
            type: string
            enum:
              - application/json
              - application/x-yaml
            default: application/json
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
            application/x-yaml:
              schema:
                $ref: "#/components/schemas/Config"
    put:
      tags:
        - Configs
      summary: Update existing phenix config
      description: >
        Can also be used to rename a config by changing the name in the metadata
        portion of the config header and using the old name in the request path.
      operationId: putConfigsKindName
      parameters:
        - name: kind
          in: path
          description: kind of phenix config to get
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: name of phenix config to get
          required: true
          schema:
            type: string
      requestBody:
        description: phenix config update parameters
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Config"
          application/x-yaml:
            schema:
              $ref: "#/components/schemas/Config"
          multipart/form-data:
            schema:
              type: object
              properties:
                fileupload:
                  type: string
                  format: binary
      responses:
        "204":
          description: successful config update
          headers:
            Location:
              schema:
                description: location of updated config
                type: string
                format: uri
    delete:
      tags:
        - Configs
      summary: Delete existing phenix config
      description: ""
      operationId: deleteConfigsKindName
      parameters:
        - name: kind
          in: path
          description: kind of phenix config to delete
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: name of phenix config to delete
          required: true
          schema:
            type: string
      responses:
        "204":
          description: successful operation
  "/schemas/{version}":
    get:
      tags:
        - Schemas
      summary: Get full schema (all config specs) for specific version
      description: ""
      operationId: getSchemasVersion
      parameters:
        - name: version
          in: path
          description: version of phenix config schema to get
          required: true
          schema:
            type: string
            enum:
              - "v1"
              - "v2"
        - name: Accept
          in: header
          description: content format for response
          required: false
          schema:
            type: string
            enum:
              - application/json
              - application/x-yaml
            default: application/json
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            application/x-yaml:
              schema:
                type: object
                additionalProperties: true
  "/schemas/{kind}/{version}":
    get:
      tags:
        - Schemas
      summary: Get schema for specific config spec kind/version
      description: ""
      operationId: getSchemasKindVersion
      parameters:
        - name: kind
          in: path
          description: kind of phenix config schema to get
          required: true
          schema:
            type: string
            enum:
              - topology
              - scenario
              - experiment
              - image
              - user
              - role
        - name: version
          in: path
          description: version of phenix config schema to get
          required: true
          schema:
            type: string
            enum:
              - "v1"
              - "v2"
        - name: Accept
          in: header
          description: content format for response
          required: false
          schema:
            type: string
            enum:
              - application/json
              - application/x-yaml
            default: application/json
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            application/x-yaml:
              schema:
                type: object
                additionalProperties: true
  "/experiments":
    get:
      tags:
        - Experiments
      summary: Get all phenix experiments
      description: ""
      operationId: getExperiments
      parameters:
        - name: screenshot
          in: query
          description: whether or not to include VM screenshots (value is size to grab)
          required: false
          schema:
            type: string
        - name: vms
          in: query
          description: >-
            set to false to leave each experiment's VMs out (`vms` is empty)
            and not ask minimega about them; `vm_count` and `delayed_vms` come
            from the topology. Ignored when screenshot is set.
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiments"
    post:
      tags:
        - Experiments
      summary: Create new phenix experiment
      description: ""
      operationId: postExperiments
      requestBody:
        description: phenix experiment creation parameters
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - topology
              properties:
                name:
                  type: string
                topology:
                  type: string
                scenario:
                  type: string
                disabled_apps:
                  description: scenario apps to leave out of the experiment
                  type: array
                  items:
                    type: string
                vlan_min:
                  type: integer
                vlan_max:
                  type: integer
                workflow_branch:
                  description: >-
                    stored as the phenix.workflow/branch experiment annotation,
                    so `annotations` may not also set it
                  type: string
                deploy_mode:
                  type: string
                  enum: ["", no-headnode, only-headnode, all]
                default_bridge:
                  type: string
                use_gre_mesh:
                  type: boolean
                node_annotations:
                  description: >-
                    added to every VM in the experiment's copy of the topology
                    (the topology config is unchanged); a VM's own annotation
                    for the same key wins. Values keep their JSON type, and
                    phenix/default-apps must be a boolean,
                    phenix/startup-autotunnel a list of strings, and
                    vrouter/vyos-password and vrouter/enable-ssh strings.
                  type: object
                  additionalProperties: {}
                annotations:
                  description: >-
                    added to the experiment's metadata annotations; topology
                    and scenario are set by phenix and rejected
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "204":
          description: >-
            experiment created; it is announced to websocket clients as an
            experiment create event
        "400":
          description: invalid request, such as an annotation phenix rejects
        "403":
          description: forbidden
        "409":
          description: phenix is already acting on an experiment with this name
...
  "/experiments/{name}/start":
    post:
      tags:
        - Experiments
      summary: Start existing phenix experiment
      description: ""
      operationId: postExperimentsNameStart
      parameters:
        - name: name
          in: path
          description: name of phenix experiment to start
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
  "/experiments/{name}/stop":
    post:
      tags:
        - Experiments
      summary: Stop existing phenix experiment
      description: ""
      operationId: postExperimentsNameStop
      parameters:
        - name: name
          in: path
          description: name of phenix experiment to stop
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
  "/experiments/{name}/schedule":
    get:
      tags:
        - Experiments
      summary: Get current schedule for existing experiment
      description: ""
      operationId: getExperimentsNameSchedule
      parameters:
        - name: name
          in: path
          description: name of phenix experiment to get schedule for
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    post:
      tags:
        - Experiments
      summary: Update current schedule for existing experiment
      description: ""
      operationId: postExperimentsNameSchedule
      parameters:
        - name: name
          in: path
          description: name of phenix experiment to update schedule for
          required: true
          schema:
            type: string
      requestBody:
        description: scheduling algorithm to use
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                algorithm:
                  type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
...
servers:
  - url: http://localhost:3000/api/v1
components:
  schemas:
...
    Configs:
      type: object
      properties:
        configs:
          type: array
          items:
            $ref: "#/components/schemas/Config"
    Config:
      type: object
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            annotations:
              type: object
              additionalProperties:
                type: string
          required:
            - name
        spec:
          type: object
          additionalProperties: true
      required:
        - apiVersion
        - kind
        - metadata
        - spec
    Experiments:
      type: object
      properties:
        experiments:
          type: array
          items:
            $ref: "#/components/schemas/Experiment"
...
    Schedule:
      type: object
      properties:
        schedule:
          type: array
          items:
            type: object
            properties:
              vm:
                type: string
              host:
                type: string
              auto_assigned:
                type: boolean
...