# src/go/web/server.go
package web

import (
	"context"
	"fmt"
	"io/fs"
	"net"
	"net/http"
	"os"
	"strings"

	"github.com/gorilla/mux"

	"phenix/util/common"
	"phenix/util/plog"
	"phenix/web/broker"
	"phenix/web/forward"
	"phenix/web/middleware"
	"phenix/web/rbac"
	"phenix/web/scorch"
	"phenix/web/weberror"
)

type route struct {
	path    string
	handler http.Handler
	methods []string
}

const (
	resourceUsers = "users"
	verbGet       = "get"
	verbPatch     = "patch"
)

var o serverOptions //nolint:gochecknoglobals // global options

func ConfigureUsers(users []string) error {
	setUserRole := func(user *rbac.User, rname string, resources ...string) {
		if role, err := rbac.RoleFromConfig(rname); err == nil {
			_ = role.SetResourceNames(resources...)

			// allow user to get their own user details
			role.AddPolicy(
				[]string{resourceUsers},
				[]string{user.Username()},
				[]string{verbGet},
			)

			_ = user.SetRole(role)
		} else {
			plog.Error(
				plog.TypeSecurity,
				"getting role for user",
				"user",
				user.Username(),
				"role",
				rname,
				"err",
				err,
			)
		}
	}

	for _, u := range users {
		creds := strings.Split(u, ":")
		uname := creds[0]
		pword := creds[1]
		rname := creds[2]

		// User already exists.
		// Confirm existing user has specified role and update if necessary.
		if user, err := rbac.GetUser(uname); err == nil {
			if user.RoleName() != rname {
				plog.Info(
					plog.TypeSecurity,
					"updating role for existing user",
					"user",
					user.Username(),
					"old",
					user.RoleName(),
					"new",
					rname,
				)

				setUserRole(user, rname, creds[3:]...)
			}

			continue
		}

		plog.Info(plog.TypeSecurity, "creating default user", "user", uname, "role", rname)

		user := rbac.NewUser(uname, pword)

		setUserRole(user, rname, creds[3:]...)
	}

	return nil
}

//nolint:funlen // server startup
func Start(opts ...ServerOption) error {
	o = newServerOptions(opts...)
	fileServerEndpoint, err := normalizeFileServerEndpoint(o.fileServerEndpoint)
	if err != nil {
		return fmt.Errorf("invalid ui.file-server-endpoint: %w", err)
	}

	if fileServerEndpoint != "" {
		o.features["file-server"] = true
	}

	_ = ConfigureUsers(o.users)

	var (
		assets    http.FileSystem
		publicDir fs.FS // the same files as assets
	)

	if o.unbundled {
		assets = http.Dir("web/public")
		publicDir = os.DirFS("web/public")

		plog.Info(plog.TypeSystem, "serving unbundled assets")
	} else {
		var err error
		assets, err = GetAssets()
		if err != nil {
			return err
		}

		if publicDir, err = fs.Sub(publicFS, "public"); err != nil {
			return fmt.Errorf("opening embedded assets: %w", err)
		}
	}

	router, err := newRouter(assets, publicDir)
	if err != nil {
		return err
	}

	plog.Info(plog.TypeSystem, "starting websockets broker")

	go broker.Start()

	plog.Info(plog.TypeSystem, "starting scorch processors")

	scorch.Start(o.basePath)

	plog.Info(plog.TypeSystem, "starting log publisher")

	go SyncMinimegaLogs(context.Background(), o.minimegaLogs)

	plog.Info(plog.TypeSystem, "starting disk image watcher")

	WatchDisks(context.Background())

	plog.Info(plog.TypeSystem, "using base path", "path", o.basePath)
	plog.Info(plog.TypeSystem, "using JWT lifetime", "lifetime", o.jwtLifetime)

	if fileServerEndpoint != "" {
		go StartFileServer(fileServerEndpoint, o.jwtKey, o.proxyAuthHeader)
	}

	if common.UnixSocket != "" {
		var (
			router = mux.NewRouter().StrictSlash(true)
			api    = router.PathPrefix("/api/v1").Subrouter()
		)

		addRoutesToRouter(api, socketRoutes()...)

		api.Use(middleware.NoAuth)

		_ = os.Remove(common.UnixSocket)

		plog.Info(plog.TypeSystem, "starting Unix socket server", "path", common.UnixSocket)

		server := http.Server{Handler: router} //nolint:gosec // Potential Slowloris Attack

		//nolint:exhaustruct // partial initialization
		listener, err := (&net.ListenConfig{}).Listen(context.Background(), "unix", common.UnixSocket)
		if err != nil {
			return err
		}

		if o.unixSocketGID != -1 {
			plog.Info(
				plog.TypeSystem,
				"setting Unix socket group permissions",
				"gid",
				o.unixSocketGID,
			)

			err = os.Chown(common.UnixSocket, -1, o.unixSocketGID)
			if err != nil {
				return err
			}

			err := os.Chmod(common.UnixSocket, 0o770) //nolint:gosec // unix socket permissions
			if err != nil {
				return err
			}
		}

		go func() {
			err := server.Serve(listener)
			if err != nil {
				plog.Error(plog.TypeSystem, "serving Unix socket", "err", err)
			}
		}()
	}

	plog.Info(plog.TypeSecurity, "starting server", "tls_enabled", o.tlsEnabled())

	if o.tlsEnabled() {
		plog.Info(plog.TypeSystem, "starting HTTPS server", "endpoint", o.endpoint)

		//nolint:gosec // Use of net/http serve function that has no support for setting timeouts
		return http.ListenAndServeTLS(
			o.endpoint,
			o.tlsCrtPath,
			o.tlsKeyPath,
			router,
		)
	} else {
		plog.Info(plog.TypeSystem, "Starting HTTP server", "endpoint", o.endpoint)
		//nolint:gosec // Use of net/http serve function that has no support for setting timeouts
		return http.ListenAndServe(
			o.endpoint,
			router,
		)
	}
}

// newRouter routes the UI's pages and assets, and the REST API behind its
// middleware, as the server options say.
//
//nolint:funlen,maintidx // the route table
func newRouter(assets http.FileSystem, publicDir fs.FS) (*mux.Router, error) {
	router := mux.NewRouter().StrictSlash(true)

	if o.featured("tunneler-download") {
		plog.Info(plog.TypeSystem, "Serving phēnix tunneler downloads")
		router.HandleFunc("/downloads/tunneler", forward.ListTunnelers).Methods("GET")
		router.HandleFunc("/downloads/tunneler/{name}", forward.GetTunneler).Methods("GET")
	}

	if o.featured("webshark") {
		plog.Info(plog.TypeSystem, "Serving WebShark", "dir", o.websharkDir)
		router.HandleFunc("/webshark/json", GetWebSharkJSON).Methods("GET")
		router.HandleFunc("/webshark/captures/{name}", GetWebSharkCaptureFile).Methods("GET", "HEAD")
		router.HandleFunc("/webshark/watch", GetWebSharkWatch).Methods("GET")
		router.PathPrefix("/webshark/").Handler(WebSharkHandler(o.websharkDir, o.basePath))
	} else {
		// rather than the phēnix UI, which would load itself in its own frame
		router.PathPrefix("/webshark/").Handler(
			http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
				http.Error(w, "WebShark is not installed on this phēnix server", http.StatusNotFound)
			}),
		)
	}

	router.HandleFunc("/features", GetFeatures).Methods("GET")
	router.HandleFunc("/version", GetVersion).Methods("GET")
	router.HandleFunc("/builder", GetBuilder).Methods("GET")
	router.HandleFunc("/builder/save", SaveBuilderTopology).Methods("POST")

	plog.Info(plog.TypeSystem, "setting up assets")

	router.PathPrefix("/docs/").Handler(
		StaticHandler(assets, false),
	)

	router.PathPrefix("/novnc/").Handler(
		StaticHandler(assets, false),
	)

	router.PathPrefix("/xterm.js/").Handler(
		StaticHandler(assets, false),
	)

	router.PathPrefix("/assets/").Handler(
		StaticHandler(assets, true),
	)

	grapheditor, err := fs.Sub(publicDir, "grapheditor")
	if err != nil {
		return nil, fmt.Errorf("opening grapheditor assets: %w", err)
	}

	router.Handle("/grapheditor/builder.bundle.js", BuilderBundleHandler(grapheditor, o.unbundled))
	router.PathPrefix("/grapheditor/").Handler(
		StaticHandler(assets, false),
	)

	router.Handle("/favicon.ico", StaticHandler(assets, false))

	router.NotFoundHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		plog.Warn(
			plog.TypeSystem,
			"Unknown route requested",
			"route",
			r.RequestURI,
			"method",
			r.Method,
		)

		// index.html names the current build's hashed chunks, so browsers must
		// revalidate it rather than reuse a stale copy
		w.Header().Set("Cache-Control", "no-cache")

		if o.unbundled {
			http.ServeFile(w, r, "web/public/index.html")
		} else {
			f, err := assets.Open("index.html")
			if err != nil {
				plog.Error(plog.TypeSystem, "opening index.html from assets", "err", err)
				http.NotFound(w, r)
				return
			}
			defer f.Close()

			fi, err := f.Stat()
			if err != nil {
				plog.Error(plog.TypeSystem, "statting index.html from assets", "err", err)
				http.NotFound(w, r)
				return
			}

			http.ServeContent(w, r, "index.html", fi.ModTime(), f)
		}
	})

	api := router.PathPrefix("/api/v1").Subrouter()

	// OPTIONS method needed for CORS
	api.Handle("/builder/topologies", weberror.ErrorHandler(GetBuilderTopologies)).
		Methods("GET", "OPTIONS")
	api.Handle("/builder/topologies", weberror.ErrorHandler(CreateBuilderTopology)).
		Methods("POST", "OPTIONS")
	api.Handle("/builder/topologies/{name}", weberror.ErrorHandler(GetBuilderTopology)).
		Methods("GET", "OPTIONS")
	api.Handle("/builder/topologies/{name}", weberror.ErrorHandler(UpdateBuilderTopology)).
		Methods("PUT", "OPTIONS")
	api.Handle("/configs", weberror.ErrorHandler(GetConfigs)).Methods("GET", "OPTIONS")
	api.Handle("/configs", weberror.ErrorHandler(CreateConfig)).Methods("POST", "OPTIONS")
	api.Handle("/configs/{kind}/{name}", weberror.ErrorHandler(GetConfig)).Methods("GET", "OPTIONS")
	api.Handle("/configs/{kind}/{name}", weberror.ErrorHandler(UpdateConfig)).
		Methods("PUT", "OPTIONS")
	api.Handle("/configs/{kind}/{name}", weberror.ErrorHandler(DeleteConfig)).
		Methods("DELETE", "OPTIONS")
	api.Handle("/configs/download", weberror.ErrorHandler(DownloadConfigs)).
		Methods("POST", "OPTIONS")
	api.Handle("/schemas/{version}", weberror.ErrorHandler(GetSchemaSpec)).Methods("GET", "OPTIONS")
	api.Handle("/schemas/{kind}/{version}", weberror.ErrorHandler(GetSchema)).
		Methods("GET", "OPTIONS")

	api.HandleFunc("/experiments", GetExperiments).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments", CreateExperiment).Methods("POST", "OPTIONS")
	api.Handle("/experiments/builder", weberror.ErrorHandler(CreateExperimentFromBuilder)).
		Methods("POST", "OPTIONS")
	api.Handle("/experiments/builder", weberror.ErrorHandler(UpdateExperimentFromBuilder)).
		Methods("PUT", "OPTIONS")
	api.Handle("/experiments/{name}", weberror.ErrorHandler(GetExperiment)).
		Methods("GET", "OPTIONS")
	api.Handle("/experiments/{name}", weberror.ErrorHandler(UpdateExperiment)).
		Methods("PATCH", "OPTIONS")
	api.HandleFunc("/experiments/{name}", DeleteExperiment).Methods("DELETE", "OPTIONS")
	api.Handle("/experiments/{name}/apps", weberror.ErrorHandler(GetExperimentApps)).
		Methods("GET", "OPTIONS")
	api.Handle("/experiments/{name}/start", weberror.ErrorHandler(StartExperiment)).
		Methods("POST", "OPTIONS")
	api.Handle("/experiments/{name}/stop", weberror.ErrorHandler(StopExperiment)).
		Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/netflow", GetNetflow).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/netflow", StartNetflow).Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/netflow", StopNetflow).Methods("DELETE", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/netflow/ws", GetNetflowWebSocket).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{name}/topology", GetExperimentTopology).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{name}/topology/search", SearchExperimentTopology).
		Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{name}/trigger", TriggerExperimentApps).Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{name}/trigger", CancelTriggeredExperimentApps).
		Methods("DELETE", "OPTIONS")
	api.HandleFunc("/experiments/{name}/schedule", GetExperimentSchedule).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{name}/schedule", ScheduleExperiment).Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{name}/captures", GetExperimentCaptures).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/captureSubnet", StartCaptureSubnet).
		Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/stopCaptureSubnet", StopCaptureSubnet).
		Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/webshark", CreateWebSharkCapture).Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{name}/files", GetExperimentFiles).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{name}/files/download", DownloadExperimentFiles).
		Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{name}/files/delete", DeleteExperimentFiles).
		Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{name}/files/{filename}", GetExperimentFile).
		Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{name}/files/{filename}", DeleteExperimentFile).
		Methods("DELETE", "OPTIONS")
	scorch.RegisterRoutes(api)
	api.HandleFunc("/experiments/{name}/soh", GetExperimentSoH).Methods("GET", "OPTIONS")
	api.HandleFunc("/soh", GetSoHSummary).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms", GetVMs).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms", UpdateVMs).Methods("PATCH", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}", GetVM).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}", UpdateVM).Methods("PATCH", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}", DeleteVM).Methods("DELETE", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/reset", ResetVM).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/restart", RestartVM).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/start", StartVM).Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/stop", StopVM).Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/shutdown", ShutdownVM).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/redeploy", RedeployVM).Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/cdrom", ChangeOpticalDisc).
		Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/cdrom", EjectOpticalDisc).
		Methods("DELETE", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/screenshot.png", GetScreenshot).
		Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/vnc", GetVNC).Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/vnc/ws", GetVNCWebSocket).
		Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/captures", GetVMCaptures).
		Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/captures", StartVMCapture).
		Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/captures", StopVMCaptures).
		Methods("DELETE", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/captures/{iface}/stream", GetVMCaptureStream).
		Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/snapshots", GetVMSnapshots).
		Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/snapshots", SnapshotVM).Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/snapshots/{snapshot}", RestoreVM).
		Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/commit", CommitVM).Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/memorySnapshot", CreateVMMemorySnapshot).
		Methods("POST", "OPTIONS")

	api.HandleFunc("/experiments/{exp}/vms/{name}/forwards", forward.GetPortForwards).
		Methods("GET", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/forwards", forward.CreatePortForward).
		Methods("POST", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/forwards", forward.DeletePortForward).
		Methods("DELETE", "OPTIONS")
	api.HandleFunc("/experiments/{exp}/vms/{name}/forwards/{host}/{port}/ws", forward.GetPortForwardWebSocket).
		Methods("GET", "OPTIONS")

	if o.featured("vm-mount") {
		api.HandleFunc("/experiments/{exp}/vms/{name}/mount", MountVM).Methods("POST", "OPTIONS")
		api.HandleFunc("/experiments/{exp}/vms/{name}/unmount", UnmountVM).
			Methods("DELETE", "OPTIONS")
		api.HandleFunc("/experiments/{exp}/vms/{name}/files", GetMountFiles).
			Methods("GET", "OPTIONS").
			Queries("path", "{path}")
		api.HandleFunc("/experiments/{exp}/vms/{name}/files/download", DownloadMountFile).
			Methods("GET", "OPTIONS").
			Queries("path", "{path}")
		api.HandleFunc("/experiments/{exp}/vms/{name}/files/upload", UploadMountFile).
			Methods("PUT", "OPTIONS").
			Queries("path", "{path}")
		api.HandleFunc("/experiments/{exp}/vms/{name}/files/copy", CopyExperimentFileToMount).
			Methods("POST", "OPTIONS")
	}

	api.HandleFunc("/disks", GetDisks).Methods("GET", "OPTIONS")
	api.HandleFunc("/disks/snapshot", SnapshotDisk).
		Methods("POST", "OPTIONS").
		Queries("disk", "{disk}", "new", "{new}")
	api.HandleFunc("/disks/rebase", RebaseDisk).
		Methods("POST", "OPTIONS").
		Queries("disk", "{disk}", "backing", "{backing}", "unsafe", "{unsafe}")
	api.HandleFunc("/disks/resize", ResizeDisk).
		Methods("POST", "OPTIONS").
		Queries("disk", "{disk}", "size", "{size}")
	api.HandleFunc("/disks/commit", CommitDisk).Methods("POST", "OPTIONS").Queries("disk", "{disk}")
	api.HandleFunc("/disks/clone", CloneDisk).
		Methods("POST", "OPTIONS").
		Queries("disk", "{disk}", "new", "{new}")
	api.HandleFunc("/disks", DeleteDisk).Methods("DELETE", "OPTIONS").Queries("disk", "{disk}")
	api.HandleFunc("/disks/rename", RenameDisk).
		Methods("POST", "OPTIONS").
		Queries("disk", "{disk}", "new", "{new}")
	api.HandleFunc("/disks", UploadDisk).Methods("POST", "OPTIONS")
	api.HandleFunc("/disks/download", DownloadDisk).
		Methods("GET", "OPTIONS").
		Queries("disk", "{disk}")

	api.HandleFunc("/vms", GetAllVMs).Methods("GET", "OPTIONS")
	api.HandleFunc("/applications", GetApplications).Methods("GET", "OPTIONS")
	api.HandleFunc("/topologies", GetTopologies).Methods("GET", "OPTIONS")
	api.HandleFunc("/topologies/{topo}/scenarios", GetScenarios).Methods("GET", "OPTIONS")
	api.HandleFunc("/hosts", GetClusterHosts).Methods("GET", "OPTIONS")
	api.HandleFunc("/users", GetUsers).Methods("GET", "OPTIONS")
	api.HandleFunc("/users", CreateUser).Methods("POST", "OPTIONS")
	api.HandleFunc("/users/{username}", GetUser).Methods("GET", "OPTIONS")
	api.HandleFunc("/users/{username}", UpdateUser).Methods("PATCH", "OPTIONS")
	api.HandleFunc("/users/{username}", DeleteUser).Methods("DELETE", "OPTIONS")
	api.HandleFunc("/users/{username}/tokens", CreateUserToken).Methods("POST", "OPTIONS")
	api.HandleFunc("/roles", GetRoles).Methods("GET", "OPTIONS")
	api.HandleFunc("/signup", Signup).Methods("POST", "OPTIONS")
	api.HandleFunc("/login", Login).Methods("GET", "POST", "OPTIONS")
	api.HandleFunc("/logout", Logout).Methods("GET", "OPTIONS")
	api.HandleFunc("/logs", GetLogs).Methods("GET", "OPTIONS")
	api.HandleFunc("/ws", broker.ServeWS).Methods("GET")
	api.HandleFunc("/console", CreateConsole).Methods("POST", "OPTIONS")
	api.HandleFunc("/console/{pid}", GetConsole).Methods("GET", "OPTIONS")
	api.HandleFunc("/console/{pid}", DeleteConsole).Methods("DELETE", "OPTIONS")
	api.HandleFunc("/console/{pid}/ws", WsConsole).Methods("GET", "OPTIONS")
	api.HandleFunc("/console/{pid}/size", ResizeConsole).
		Methods("POST", "OPTIONS").
		Queries("cols", "{cols:[0-9]+}", "rows", "{rows:[0-9]+}")

	api.HandleFunc("/settings", GetSettings).Methods("GET", "OPTIONS")
	api.HandleFunc("/settings", SetSettings).Methods("POST", "OPTIONS")
	api.HandleFunc("/settings/password", GetPasswordRequirements).Methods("GET", "OPTIONS")
	api.HandleFunc("/settings/timeout", GetTimeoutSettings).Methods("GET", "OPTIONS")

	addRoutesToRouter(api, socketRoutes()...)

	// outermost, so the other middleware (full logging in particular) sees the
	// uncompressed response
	api.Use(CompressResponses)

	if o.allowCORS {
		plog.Info(plog.TypeSystem, "CORS is enabled on HTTP API endpoints")
		api.Use(middleware.AllowCORS)
	}

	api.Use(middleware.Auth(o.jwtKey, o.proxyAuthHeader))

	switch o.logMiddleware {
	case "full":
		plog.Info(plog.TypeSystem, "full HTTP logging is enabled")
		api.Use(middleware.LogFull)
	case "requests":
		plog.Info(plog.TypeSystem, "requests-only HTTP logging is enabled")
		api.Use(middleware.LogRequests)
	}

	return router, nil
}

// socketRoutes are the API routes the Unix socket serves as well.
func socketRoutes() []route {
	return []route{
		{"/workflow/apply/{branch}", weberror.ErrorHandler(ApplyWorkflow), []string{"POST"}},
		{
			"/workflow/configs/{branch}",
			weberror.ErrorHandler(WorkflowUpsertConfig),
			[]string{"POST"},
		},
		{"/options", weberror.ErrorHandler(GetOptions), []string{"GET"}},
	}
}

func addRoutesToRouter(router *mux.Router, routes ...route) {
	for _, r := range routes {
		router.Handle(r.path, r.handler).Methods(r.methods...)
	}
}